+++ src/components/Navigation.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+import "./components.css";
+
+export default function Navigation() {
+  return (
+    <nav>
+      <a href="#">Corrie Stroup</a>
+      <ul>
+        <li>
+          <a href="#">About</a>
+        </li>
+        <li>
+          <a href="#">Recent Work</a>
+        </li>
+        <li>
+          <a href="#">Services</a>
+        </li>
+        <li>
+          <a href="#">Contact</a>
+        </li>
+      </ul>
+    </nav>
+  );
+}
