+++ src/components/components.css
@@ -1,208 +1,17 @@
-/** SITE HEADER **/
-
-#site-header {
+nav {
+  position: absolute;
+  top: 0;
   display: flex;
   width: 100%;
-  /* background-color: red; */
-  justify-content: space-between;
-  position: relative;
-  z-index: 50;
-  background-color: var(--navy);
-  padding: var(--variable-padding-small);
-  box-sizing: border-box;
-}
+  background: black;
+  padding: 1em;
 
-#profile-picture {
-  width: 75px;
-  height: 75px;
-  object-fit: cover;
-  transition: var(--scale-out);
-  cursor: pointer;
-}
-
-#profile-picture:hover {
-  transform: scale(1.1);
-  transition: var(--scale-in);
-}
-
-#connect-box {
-  display: flex;
-  align-items: center;
-}
-
-#connect-box img {
-  filter: invert(75%);
-  transition: var(--scale-out);
-}
-
-#connect-box img:hover {
-  transition: var(--scale-in);
-  transform: scale(1.1);
-  filter: invert(100%);
-}
-
-.connect-icon {
-  width: 50px;
-  height: 50px;
-}
-
-#blog-icon {
-  width: 40px;
-  height: 40px;
-  margin: 5px;
-  margin-top: 0;
-}
-
-/** BLURB DISPLAY **/
-#blurb-container {
-  z-index: 50;
-  position: relative;
-  background-color: var(--navy);
-  padding: var(--padding-large);
-  height: fit-content;
-}
-
-#blurb-container .highlight {
-  color: var(--light-green);
-  font-weight: bold;
-}
-
-@media only screen and (min-width: 700px) {
-  #blurb-container {
-    grid-column: 1;
+  ul {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+    display: flex;
+    margin: auto;
+    gap: 1.5em;
   }
-}
-
-.project-blurb {
-  margin-top: 1em;
-  border-left: 3px solid var(--light-green);
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-  padding-left: var(--padding-large);
-  font-family: var(--font-main);
-  color: var(--light-slate);
-}
-
-.project-blurb a {
-  color: var(--light-green);
-}
-
-.project-img img {
-  max-width: 100%;
-  object-fit: contain;
-  margin: auto;
-}
-
-.hidden {
-  transform: scaleX(0);
-  opacity: 0;
-  position: absolute;
-}
-
-.blurb {
-  transform-origin: right;
-  height: fit-content;
-  transition: opacity 0s, transform 0.5s;
-}
-
-.project-links {
-  display: flex;
-  justify-content: center;
-  gap: 2em;
-  color: var(--light-green);
-}
-
-.blurb-transition {
-  transition: 0.5s;
-}
-
-/** PROJECT MENU **/
-#projects-label {
-}
-
-#project-menu-container {
-  z-index: 50;
-  background: var(--navy);
-  width: 100%;
-  height: fit-content;
-  position: relative;
-  margin-top: -1px;
-  padding: var(--padding-large);
-  box-sizing: border-box;
-  flex-shrink: 1;
-}
-
-@media only screen and (min-width: 700px) {
-  #project-menu-container {
-    /* width: 185px; */
-    width: 1fr;
-    grid-column: 2;
-    width: 20em;
-  }
-}
-
-#project-menu {
-  margin-top: 2em;
-  padding: var(--padding-large);
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-  width: fit-content;
-  list-style: none;
-  position: relative;
-  width: max-content;
-  box-sizing: border-box;
-}
-
-#project-menu::before {
-  position: absolute;
-  width: 33px;
-  height: 100%;
-  border-top-left-radius: var(--border-radius);
-  content: "";
-  left: 0;
-  top: 0;
-  border-left: 3px solid var(--light-green);
-  border-top: 3px solid var(--light-green);
-}
-
-#project-menu a {
-  box-sizing: border-box;
-  transition: 0.75s;
-  width: fit-content;
-}
-
-#project-menu a:hover {
-  letter-spacing: 1px;
-  cursor: pointer;
-  transition: 0.3s;
-}
-
-/** BACKGROUND LAYER **/
-
-#background-layer-top,
-#background-layer-bottom {
-  background: var(--light-green);
-  z-index: 10;
-  width: 400px;
-  height: 800px;
-  position: absolute;
-  overflow: hidden;
-}
-
-#background-layer-top {
-  top: 0;
-  left: 0;
-  /* border-radius: 50%; */
-  border-top-left-radius: var(--border-radius);
-  transform: translate(-50%, -50%) rotate(45deg);
-}
-
-#background-layer-bottom {
-  bottom: 0;
-  right: 0;
-  border-radius: 50%;
-  border-top-left-radius: var(--border-radius);
-  transform: translate(50%, 50%) rotate(45deg);
 }
