+++ src/index.css
@@ -66,7 +66,7 @@ body,
 #root {
   display: flex;
   justify-content: center;
-  align-items: center;
+  /* align-items: center; */
   background-color: var(--navy);
   height: fit-content;
   /* overflow: hidden; */
