+++ src/sections/Introduction.jsx
@@ -1,0 +1,26 @@
+import React from "react";
+import "./sections.css";
+export default function Introduction() {
+  return (
+    <div id="introduction" className="section">
+      <h1>Greetings!</h1>
+      <p>
+        My name is <span className="highlight">Corrie Stroup</span>, I'm a
+        Frontend Developer based in Ho Chi Minh City, Vietnam. I have developed
+        a variety of different apps from reproductions of popular social media
+        websites to business UIs.
+      </p>
+      <br />
+      <p>
+        <span className="highlight">Expressive, creative front-ends</span> are a
+        passion of mine. In a world increasingly technological, cutting-edge is
+        where I like to be; the frontier of pixel-perfection and intuitive UX.
+      </p>
+      <br />
+      <p>
+        New stacks, old stacks, all stacks have a time and a place and provide
+        value in development.
+      </p>
+    </div>
+  );
+}
