+++ src/sections/sections.css
@@ -1,0 +1,3 @@
+.section {
+  margin-top: 5em;
+}
